Change to suggested directory structure in review
