+++ lms/services/canvas_api/_authenticated.py
